# apps/inventory/openapi.yaml
openapi: 3.0.0
info:
  title: Inventory Service API
  version: 1.0.0
  description: 食材の在庫管理を行うAPI
servers:
  - url: http://localhost:3001
paths:
  /api/ingredients:
    get:
      summary: 食材一覧取得
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
    post:
      summary: 食材登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - quantity
                - unit
              properties:
                name:
                  type: string
                quantity:
                  type: number
                unit:
                  type: string
                category:
                  type: string
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
  /api/ingredients/{id}:
    put:
      summary: 食材更新
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientInput'
      responses:
        '200':
          description: 更新成功
    delete:
      summary: 食材削除
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 削除成功

components:
  schemas:
    Ingredient:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    IngredientInput:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        category:
          type: string